+++ src/main/java/org/opensearch/neuralsearch/mapper/dto/SemanticParameters.java
@@ -7,6 +7,8 @@ package org.opensearch.neuralsearch.mapper.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.util.Map;
+
 /**
  * A DTO to hold all the semantic parameters.
  */
@@ -19,4 +21,5 @@ public class SemanticParameters {
     private final String semanticInfoFieldName;
     private final Boolean chunkingEnabled;
     private final String semanticFieldSearchAnalyzer;
+    private final Map<String, Object> denseEmbeddingConfig;
 }